refactor(agent_common): de-duplicate shared callback + schema, drop dead config
- hoist the byte-identical rate_limit_callback into
  agent_common/rate_limit.py as build_rate_limit_callback(config); both
  creative_agent and trend_scout now import it (per-agent rpm knobs preserved)
- consolidate the field-identical SearchQuery model into creative_agent.schemas;
  both research sub-agents import it (resolves their # TODO: consolidate)
- remove the dead SetupConfiguration class + setup_config instance from
  trend_scout/config.py (grep-confirmed unreferenced)
