+++ creative_agent/callbacks.py
@@ -2,16 +2,15 @@ import uuid
 import warnings
 import pandas as pd
 import re
-import time
 import logging
 from typing import Optional, Dict, Any
 
 from google.genai import types
 from google.adk.sessions.state import State
-from google.adk.models.llm_request import LlmRequest
 from google.adk.agents.callback_context import CallbackContext
 
 from agent_common import observability, sanitize
+from agent_common.rate_limit import build_rate_limit_callback
 
 from .config import config
 
@@ -85,47 +84,10 @@ def load_session_state(callback_context: CallbackContext):
     _set_initial_states(data["state"], callback_context.state)
 
 
-def rate_limit_callback(
-    callback_context: CallbackContext, llm_request: LlmRequest
-) -> None:
-    # pylint: disable=unused-argument
-    """Callback function that implements a query rate limit.
-
-    Args:
-      callback_context: A CallbackContext object representing the active
-              callback context.
-      llm_request: A LlmRequest object representing the active LLM request.
-    """
-    now = time.time()
-    if "timer_start" not in callback_context.state:
-        callback_context.state["timer_start"] = now
-        callback_context.state["request_count"] = 1
-        logging.debug(
-            "rate_limit_callback [timestamp: %i, req_count: 1, elapsed_secs: 0]",
-            now,
-        )
-        return
-
-    request_count = callback_context.state["request_count"] + 1
-    elapsed_secs = now - callback_context.state["timer_start"]
-    logging.debug(
-        "rate_limit_callback [timestamp: %i, request_count: %i, elapsed_secs: %i]",
-        now,
-        request_count,
-        elapsed_secs,
-    )
-
-    if request_count > config.rpm_quota:
-        delay = config.rate_limit_seconds - elapsed_secs + 1
-        if delay > 0:
-            logging.debug("Sleeping for %i seconds", delay)
-            time.sleep(delay)
-        callback_context.state["timer_start"] = now
-        callback_context.state["request_count"] = 1
-    else:
-        callback_context.state["request_count"] = request_count
-
-    return
+# Shared query rate limiter (agent_common). Built with creative_agent's config so
+# `callbacks.rate_limit_callback` keeps the same name/signature for the
+# before_model_callback wiring in agent.py (and interactive_creative reuse).
+rate_limit_callback = build_rate_limit_callback(config)
 
 
 def collect_research_sources_callback(callback_context: CallbackContext) -> None:
