+++ trend_scout/callbacks.py
@@ -1,4 +1,3 @@
-import time
 import uuid
 import logging
 import warnings
@@ -6,10 +5,10 @@ import pandas as pd
 from typing import Dict, Any
 
 from google.adk.sessions.state import State
-from google.adk.models.llm_request import LlmRequest
 from google.adk.agents.callback_context import CallbackContext
 
 from agent_common import observability
+from agent_common.rate_limit import build_rate_limit_callback
 
 from .config import config
 
@@ -81,44 +80,7 @@ def load_session_state(callback_context: CallbackContext):
         callback_context.state["interactive_trend_pick"] = False
 
 
-def rate_limit_callback(
-    callback_context: CallbackContext, llm_request: LlmRequest
-) -> None:
-    # pylint: disable=unused-argument
-    """Callback function that implements a query rate limit.
-
-    Args:
-      callback_context: A CallbackContext object representing the active
-              callback context.
-      llm_request: A LlmRequest object representing the active LLM request.
-    """
-    now = time.time()
-    if "timer_start" not in callback_context.state:
-        callback_context.state["timer_start"] = now
-        callback_context.state["request_count"] = 1
-        logging.debug(
-            "rate_limit_callback [timestamp: %i, req_count: 1, elapsed_secs: 0]",
-            now,
-        )
-        return
-
-    request_count = callback_context.state["request_count"] + 1
-    elapsed_secs = now - callback_context.state["timer_start"]
-    logging.debug(
-        "rate_limit_callback [timestamp: %i, request_count: %i, elapsed_secs: %i]",
-        now,
-        request_count,
-        elapsed_secs,
-    )
-
-    if request_count > config.rpm_quota:
-        delay = config.rate_limit_seconds - elapsed_secs + 1
-        if delay > 0:
-            logging.debug("Sleeping for %i seconds", delay)
-            time.sleep(delay)
-        callback_context.state["timer_start"] = now
-        callback_context.state["request_count"] = 1
-    else:
-        callback_context.state["request_count"] = request_count
-
-    return
+# Shared query rate limiter (agent_common). Built with trend_scout's config so
+# `callbacks.rate_limit_callback` keeps the same name/signature for the
+# before_model_callback wiring in agent.py.
+rate_limit_callback = build_rate_limit_callback(config)
